update release-related steps to work on the new deploy queue
The agents processing the "deploy" queue have changed to an elastic
stack.

They have the tools required (awscli, docker) and the required IAM
permissions (uploading to S3), but we need to read the github API from
SSM ourselves.

One of the steps also uses github-release, which isn't available on the
elastic stack. I've switched from manually running a docker container
with the binary, to running the entire step in a container that has it.
This is mainly for consistency with other pipelines that use
github-release.