Update release process to use OIDC
Now that OIDC is available, we prefer to use it to assume specific IAM
roles for deploy tasks.

For pushing to S3, assume an IAM role from the account where the S3
buckets live.

For releasing to GitHub, assume role in the account running the step and
use that to set an ENV var from SSM Parameter Store