Handle API errors when querying queue
When collecting metrics for all queues a non `http.StatusOK` response
from the API is correctly checked to ensure its content type is JSON
before attempting to JSON decode it.

Unfortunately this same check was omitted from the code path that
retrieves metrics when individual queues have been specified.

For example in the event of a 500 response ,which is returned as HTML,
failing to check the response content type leads to the JSON decoder
trying to decode HTML and the following error:
`invalid character '<' looking for beginning of value: SyntaxError`
which isn't very useful to the user.