Remove name from release candidate branch
I think it's better that the release candidate branches are NOT
namespaced by their author so that there is some kind of barrier
preventing multiple authors from releasing at the same time. Of course
this dose not replace proper communication in the organisation which
should be the primary way to ensure this does not happen.