Ignore cluster if it is empty
\#227 added support for clusters, but as was somewhat anticipated, it
broke compatibility with the some existing backends because it added a
new dimension/label/tag or some equivalent for the cluster. In this PR,
we refrain from doing so if the cluster is the empty string.

One thing to note is that the cluster is not specified when the process
starts, it is encoded in the cluster token, and the backend include the name
of the cluster in its response. If the token is for unclustered agents,
that part of the response will be the empty string.