This change enables configuration of http connection pooling
To provide customers with more options to tune for their situation this
introduces a new option to reduce or disable connection reuse. In some
cases disabling connection reuse may avoid read timeouts or pausing for
connections a server, or CDN may deprioritize.

I have also updated the documentation for the expanded timeout option to
highlight it is used for more than just request timeout.