fix to remove reset in prometheus backend
This has been reoved as it can't work for more than one cluster as gauges are scoped to all clusters, but the collect function is called once for each cluster. In summary this leads to gauges only being correct for the last cluster, depending what metrics were returned when the cluster was polled as the result doesn't always include all keys.