Simplify docker release script
It's not super clear to my why we can't do the multi-platform image
build *and* push to ECR all in a single command. It might be naive
optimism, but it's worth a shot. I don't expect this to speed things
up too much, but a simpler script will be easier to maintain over time.

I also removed the creation of a custom buildx builder. These jobs are
running on elastic-runners, which already have a buildx builder that's
capable of multi-platform builds.