Update release process to assume new IAM roles via OIDC
We recently moved the S3 buckets that host our public lambdas to a new
AWS account, and the OIDC assumable IAM roles for publishing to the
buckets moved with them.

The replacement buckets have the same name, so we don't need to change them.
However the recreated buckets use BucketOwnerEnforced for ownership and
resource policies for access, instead of ACLs. We need to remove `--acl`
from the upload command.

We also took the opportunity to setup S3 bucket replication, so we only
need to push to the us-east-1 bucket now. S3 replication will handle
copying to all the regional buckets.