Remove redundant queue scope
The default queue for the Main cluster is the `elastic-runners` queue
anyway. But the real motivator for the removal is that we're looking at
moving this pipeline to use the OSS cluster, which has a different set
of queues. The default queue is expected to be fine on the OSS cluster,
which means we can just do away with the queue scope entirely and rely
on cluster defaults.