Remove hardcoded RELEASE_DRY_RUN
Hardcoding here means that nothing can override it, which somewhat
defeats the purpose of having the environment variable in the first
place. Order of precedence from the docs [1], where values in each
successive set take precedence:

> Pipeline, Build, Step, Standard

Step is pipeline.yml. Build is creating a build via the web UI, which is
where we'd like to be able to set RELEASE_DRY_RUN=true on occasion

1. https://buildkite.com/docs/pipelines/configure/environment-variables#environment-variable-precedence